Build: Application, Resource 폴더 정리
