+++ GIL/GIL/Resources/GoogleService-Info.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyBZqFF27ChPx7kZqgUB89ZjgihPHkQiiMs</string>
+	<key>GCM_SENDER_ID</key>
+	<string>453591206459</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>woojin.GIL</string>
+	<key>PROJECT_ID</key>
+	<string>gil-ios-a685d</string>
+	<key>STORAGE_BUCKET</key>
+	<string>gil-ios-a685d.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:453591206459:ios:0b97c8fca754c742de6a1e</string>
+</dict>
+</plist>
